Enable node imports by absolute path
Node imports are now allowed from absolute paths, starting with `src/`.

For example, imports from "../../core/graph" can now be replaced with
"core/graph".

Alternatives considered: Consider that we want to import the file
`src/core/graph`. I considered the following four approaches for
this import:

1. `import {Stuff} from "core/graph"`
2. `import {Stuff} from "Core/graph"`
3. `import {Stuff} from "@core/graph"`
4. `import {Stuff} from "@/core/graph"`

I chose option 1, because it is both concise and easy to implement.

Because we don't use a separate namespace for our src/ directories and
node modules, there is a possibility of name collisions.  (If so,
node_modules will win.) However, I believe this issue is unlikely to
happen in practice, as we have only a few directories in `src`, we would
need a conflict on the full (pathed) name.

In the event that such a conflict does arise, and is not easily resolved
by disambiguating the directory name, it should not be disruptive to
switch to import aliasing at that time.

Test plan: This commit does not change any actual imports. To confirm
that it works, you can change some imports, as with the following two
commands:

```
find src -name '*.js' | xargs sed -ri 's%(from ")(../)+(plugins/)?(git|github|artifact)/%\1plugins/\4/%g'
find src -name '*.js' | xargs sed -ri 's%(from ")(../)+(core|graphql|tools|app)?/%\1\3/%g'
```