Fix sample GitHub plugin serialization
Summary:
The previous code was attempting to check `rawJson.type` before invoking
`fromCompat`. This suggests that the plugin should have a uniform compat
format… or that we simply want the type-and-version to be one
identifier, and to have handlers corresponding to types-and-versions,
not just versions—but that would be a larger change.

Test Plan:
Flow passes now that we’ve given actual types to the `compat` shim.