Reorganize Graph address-related code
Our Graph uses specialized strings as addresses. These addresses are
exported as an opaque type; they can't be manipulated via standard
string methods, but instead require special address-aware methods.
Addresses have path-like semantics (they are hierarchically orgnaized by
string components), and we expect to have methods for appending
components, finding prefixes, concatenting addresses, mapping them to
and from string arrays, etc.

It would be nice to attach these methods to the addresses, so it
would be possible to call `address.append(component)`,
`address.prefix()`, and so forth. However, as addresses are strings,
this would require modifying the string prototype. We could make
addresses classes instead, that are backed by or convertible to strings,
but this would break the ability to test address equality using `===`.

After some offline discussion with @wchargin, we decided the best
compromise is to create a "psuedo-module" which contains all address
related methods, and can be imported from the `core/graph` module. This
commit creates that module, called `Address`, which contains the
following methods:

- `nodeAddress`: The factory function for node addresses. Returns a base
`NodeAddress` if called without arguments, or constructs a componentized
address if called with an array of parts.

- `edgeAddress`: The factory function for edge addresses. Returns a base
`EdgeAddress` if called without arguments, or constructs a componentized
address if called with an array of parts.

- `toParts`: Takes a node or edge address and returns a string array of
parts.

The module is implemented in `address.js`. Clients are not intended to
import `address.js` directly; rather, `graph.js` re-exports the public
pieces. (This is just for clients' convenience; being able to import all
Graph-required functionality from one file is pleasant.)

A note about `NodeAddress` and `EdgeAddress`: There are actually two
types of Addresses (Node and Edge), which have different
representations. It is a type and runtime error to pass a `NodeAddress` to
a function expecting an `EdgeAddress` and vice versa. Offline, @wchargin
and I discussed exporting two psuedomodules, one for node and one for
edge addresses. On further consideration, I believe this introduces
needless API complexity. It will be easy to write the single `Address`
module with functions that are apprporiately generic over Node/Edge
addresses, so we will still get type (and runtime) errors when passing
(for example) an `EdgeAddress` into `Graph.hasNode`.

`address.js` also exports `assertNodeAddress` and `assertEdgeAddress`.
These functions aren't yet used, but will be needed for appropriate
runtime assertions in `Graph` methods.

Test plan: Observe unit tests, Travis.